+++ install/src/main/java/org/jetbrains/bsp/bazel/install/LauncherArgumentCreator.kt
@@ -30,4 +30,7 @@ class LauncherArgumentCreator(private val installationContext: InstallationConte
             ?.let { "-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=$it" }
 
     fun projectViewFilePathArgv(): String? = installationContext.projectViewFilePath?.toString()
+
+    fun bazelWorkspaceRootDir(): String = installationContext.bazelWorkspaceRootDir.toString()
+
 }