+++ install/src/main/java/org/jetbrains/bsp/bazel/install/cli/CliOptions.kt
@@ -31,4 +31,5 @@ data class CliOptions internal constructor(
         val projectViewFilePath: Path?,
         val projectViewCliOptions: ProjectViewCliOptions?,
         val bloopCliOptions: BloopCliOptions,
+        val bazelWorkspaceRootDir: Path,
 )