+++ install/src/main/java/org/jetbrains/bsp/bazel/install/cli/CliOptionsProvider.kt
@@ -165,6 +165,13 @@ class CliOptionsProvider(private val args: Array<String>) {
             .desc("Use bloop as the BSP server rather than bazel-bsp.")
             .build()
         cliParserOptions.addOption(useBloopOption)
+
+        val bazelWorkspaceRootDirOption = Option.builder(BAZEL_WORKSPACE_ROOT_DIR_OPT)
+            .longOpt("bazel_workspace")
+            .hasArg()
+            .desc("Add path to Bazel project's root directory. By default, it is the same as -$BAZEL_PATH_SHORT_OPT")
+            .build()
+        cliParserOptions.addOption(bazelWorkspaceRootDirOption)
     }
 
     fun getOptions(): Try<CliOptions> {
@@ -180,6 +187,7 @@ class CliOptionsProvider(private val args: Array<String>) {
             projectViewFilePath = projectViewFilePath(cmd),
             projectViewCliOptions = createProjectViewCliOptions(cmd),
             bloopCliOptions = createBloopCliOptions(cmd),
+            bazelWorkspaceRootDir = bazelWorkspaceRootDir(cmd),
         )
 
     private fun workspaceRootDir(cmd: CommandLine): Path =
@@ -193,6 +201,9 @@ class CliOptionsProvider(private val args: Array<String>) {
             isHelpOptionUsed = isHelpOptionUsed(cmd),
             printHelp = ::printHelp,
         )
+
+    private fun bazelWorkspaceRootDir(cmd: CommandLine): Path =
+        getOptionValueAndMapToAbsolutePath(cmd, BAZEL_WORKSPACE_ROOT_DIR_OPT) ?: workspaceRootDir(cmd)
 
     private fun isHelpOptionUsed(cmd: CommandLine): Boolean = cmd.hasOption(HELP_SHORT_OPT)
 
@@ -293,6 +304,7 @@ class CliOptionsProvider(private val args: Array<String>) {
         private const val DERIVE_TARGETS_FLAG_SHORT_OPT = "v"
         private const val IMPORT_DEPTH_SHORT_OPT = "i"
         private const val USE_BLOOP_SHORT_OPT = "u"
+        private const val  BAZEL_WORKSPACE_ROOT_DIR_OPT ="w"
 
         const val INSTALLER_BINARY_NAME = "bazelbsp-install"
     }